Move common methods to ci
